Moved everything to the root folder as we don't need internal.
