+++ api/handler/handler.go
@@ -1,0 +1,7 @@
+package handler
+
+import "github.com/labstack/echo/v4"
+
+type Handler interface {
+	Register(api *echo.Group)
+}
