+++ domain/model/catalog/group.go
@@ -1,0 +1,15 @@
+package catalog
+
+import (
+	"github.com/compliance-framework/configuration-service/domain/model"
+)
+
+type Group struct {
+	Uuid model.Uuid `json:"uuid"`
+
+	model.ComprehensiveDetails
+
+	Class  string       `json:"class"`
+	Params []Parameter  `json:"params"`
+	Groups []model.Uuid `json:"groups"`
+}
