+++ domain/model/link.go
