+++ domain/model/part.go
