+++ domain/model/property.go
