--- internal/domain/model/catalog/part.go
@@ -1,16 +1,0 @@
-package catalog
-
-import (
-	"github.com/compliance-framework/configuration-service/internal/domain/model"
-)
-
-type Part struct {
-	Uuid  model.Uuid       `json:"uuid"`
-	Name  string           `json:"name"`
-	Ns    string           `json:"ns"`
-	Class string           `json:"class"`
-	Title string           `json:"title"`
-	Props []model.Property `json:"props"`
-	Parts []Part           `json:"parts"`
-	Links []model.Link     `json:"links"`
-}
